Add files via upload
Source Code is uploaded in .py and .ipynb formats. Provisions for finding the crude prediction for "market cap" parameter is provided at the end of the Source Code. To get the model work correctly, please enter a date from the near future (relative to the dates the model has been trained). One can also make an advanced prediction by entering the expected value of all other parameters but the one which is to be predicted (this model takes into account the dependability of parameters onto each other. However, the drawback is that it needs the value of other params on the future date.